requirements.txt
pandas
sklearn
matplotlib
scipy
requests
beautifulsoup4
click
jinja2
IPython
openstacksdk
jupyter
